Smarter deduplication
This new deduplication mechanism uses a map between candidate pais (in
the left side of the three way join) and optional capabilities. If the
capability is None, then the pair has already been sent, otherwise we
consume the capability to send it, leaving a None in its place. This
prevents sending duplicates.