Change the adaptive algorithm
Now the configuration accepts an additional parameter: the desired size of
the buckets. This way we can control with how many other points each one
is compared with, which might not be the number minimizing the total cost.
This is because often the majority of the points are assigned the top
level, but actually we don't mind performing a few more comparisons.